api/requirements.txt
# 最小化依賴套件 - 用於快速安裝

# FastAPI 框架
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.20

# 機器學習和 AI
torch>=2.1.0
faster-whisper>=0.9.0
numpy>=1.21.0
transformers>=4.30.0

# 音訊處理py
librosa>=0.10.0
soundfile>=0.12.0
webrtcvad>=2.0.10

# 語音辨識相關
opencc-python-reimplemented>=0.1.0
cn2an>=0.5.0

# STT Streaming 核心依賴
websockets>=13.0.0

# 網路和 API
requests>=2.31.0

# 其他工具
python-dotenv==1.0.1
tqdm==4.66.5

# Auth
PyJWT>=2.8.0
passlib>=1.7.4

# Testing
pytest>=8.2.0
httpx>=0.27.0